project/Backend/requirements.txt
flask~=3.1.3
flask-cors~=6.0.2
PyJWT~=2.11.0
cryptography~=46.0.5
python-dotenv~=1.1.0
requests~=2.32.3
pymongo~=4.11.3
pdfminer.six==20260107
skillNer==1.0.3
spacy~=3.8.11
cloudinary~=1.41.0
sentence-transformers~=3.4.1
qdrant-client~=1.13.3
